requirements.txt
django==5.1.6
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.1.0
djangorestframework-simplejwt==5.4.0
